Fix null thread pointer dereference on init
When Kernel::Start() is called, the port-specific kernel runtime
initialization begins.  In this function, we typically instantiate
the KernelTimer thread.  Unfortunately, in this context, the kernel
has been marked as started (as Kernel::Start() set the kernel-started
flag before calling the port-specific code), causing the Start() call
to perform calls against the current running thread (which is not set)