Implement portable profiling timer
Implement a portable profiling timer based on the common Kernel::GetTicks()
API call, instead of requiring a part-specific implementation.  Profiling
timer is now based on a 1ms tick as a result.  As a result of this change,
all port-specific implementations (and stubs) for profiling timers have been
removed.  The unit tests are also updated to handle this change, with test
limits updated accordingly.